Rename MediaRecorderView to ChatMediaRecorderView
